move LatestWarningsDashboardWidget to warnings folder
